apps/api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
langgraph
langchain
langchain-openai
sqlglot
lxml
pydantic
networkx
pytest
pytest-asyncio
httpx
python-multipart
bcrypt>=4.0.0
cryptography>=41.0.0
